Moved get logic to requests folder
